+++ app/Strings.hs
@@ -44,7 +44,7 @@ where
 --
     version :: String
     version = unpack [text|
-        0.6.0
+        0.6.1
     |]
 
 --