+++ src/GWind/Service/Control.hs
@@ -1,23 +1,37 @@
--- ⚠
+--
 module GWind.Service.Control (reviseServicesM, createServicesM, deleteServicesM)
 where
 
     import GWind.Project.Domain (ProjectSnake)
     import GWind.Service.Domain (ServiceNicks)
+    import GWind.Project.Mutator (projectId)
+    import GWind.Service.Mutator (serviceIds)
+    import GWind.Service.Service (revisionServicesM, creationCodeM, deletionCodesM)
+    import GWind.Common.Presenter (presentingM)
 
 --
     reviseServicesM :: ProjectSnake -> IO ()
-    reviseServicesM _ =
-        undefined
+    reviseServicesM projectSnake =
+        let pid = projectId projectSnake
+        in revisionServicesM pid
+            >>= presentingM
 
 --
     createServicesM :: ProjectSnake -> ServiceNicks -> IO ()
-    createServicesM _ _ =
-        undefined
+    createServicesM projectSnake serviceNicks =
+        let
+            pid = projectId projectSnake
+            sis = serviceIds serviceNicks
+        in creationCodeM pid sis
+            >>= presentingM
 
 --
     deleteServicesM :: ProjectSnake -> ServiceNicks -> IO ()
-    deleteServicesM _ _ =
-        undefined
+    deleteServicesM projectSnake serviceNicks =
+        let
+            pid = projectId projectSnake
+            sis = serviceIds serviceNicks
+        in deletionCodesM pid sis
+            >>= presentingM
 
 --