+++ src/GWind/Service/Domain.hs
@@ -1,7 +1,8 @@
 -- ⚠
-module GWind.Service.Domain (ServiceNicks)
+module GWind.Service.Domain (ServiceNicks, ServiceId)
 where
 
     type ServiceNicks = String
+    type ServiceId = String
 
 --